abhaya_raksha/backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi==0.135.1
uvicorn==0.42.0
starlette==0.52.1
python-multipart==0.0.22

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.48

# ── Config & Env ──────────────────────────────────────────────────────────────
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2

# ── Auth — CRITICAL PINS ──────────────────────────────────────────────────────
# bcrypt 4.x removed __about__ module; passlib 1.7.4 requires it.
# auth.py contains a monkey-patch to bridge this gap. Do NOT upgrade either
# without re-testing verify_password() end-to-end.
bcrypt==4.0.1
passlib==1.7.4
# python-jose enforces JWT sub claim as string (RFC 7519).
# auth.py casts sub to str() on encode and int() on decode.
python-jose[cryptography]==3.5.0
cryptography==46.0.5

# ── HTTP Client ───────────────────────────────────────────────────────────────
httpx==0.28.1

# ── ML Stack ─────────────────────────────────────────────────────────────────
scikit-learn==1.8.0
numpy==2.4.3
pandas==3.0.1
joblib==1.5.3
scipy==1.17.1

# ── AI ────────────────────────────────────────────────────────────────────────
google-generativeai==0.8.6

# ── Scheduling ────────────────────────────────────────────────────────────────
apscheduler==3.11.2

# ── Payment Gateway ───────────────────────────────────────────────────────────
razorpay==1.4.2
